# api/openapi/qavia.yaml
openapi: 3.1.0
...
info:
  title: Qavia API
  version: 0.1.0
  description: |
    Internal Hyscaler platform that acts as an AI QA engineer.

    Every mutation goes through this API. The web app holds no database
    credentials and no queue access, so this is the single place authorisation is
    enforced.
  contact:
    name: Hyscaler
    email: max@hyscaler.com
servers:
  - url: /
    description: Same origin as the API process
...
security:
  - sessionCookie: []
paths:
...
  /api/v1/projects/{projectID}/test-data/shapes:
    get:
      operationId: listTestDataShapes
      tags: [testdata]
      summary: What this project can generate data for
      description: |
        One entry per endpoint with a schema worth generating from, with its field names
        so a UI can offer per-field AI enrichment rather than asking somebody to type a
        path exactly. Endpoints that declare no request or response schema are omitted:
        an option that produces one empty record is not an option.
      parameters:
        - $ref: "#/components/parameters/ProjectID"
      responses:
        "200":
          description: The shapes available.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestDataShapeList"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/projects/{projectID}/test-data:
    post:
      operationId: generateTestData
      tags: [testdata]
      summary: Generate test data from the specification's own schemas
      description: |
        Deterministic and free (F-10.1). The response carries the seed that produced it,
        and the same seed and count reproduce the data byte for byte — which is what
        makes a failure on record 63 reproducible instead of a story about somebody's
        machine.

        Type, format, pattern, enum, and the length and range bounds are honoured,
        because data that violates the schema tests the validator rather than the
        endpoint. **Payment card numbers come only from published test ranges** and are
        never freshly generated (F-10.4).

        `fields` opts individual fields into an AI pass for the cases where semantics
        matter — a plausible address, a correctly shaped tax identifier (F-10.2). It is
        per field on purpose: the bulk is a seeded faker, and a model asked for two
        thousand values would cost money to be occasionally wrong. Whatever the model
        does not answer keeps the generated value, and the response says so.

        Set `format` to stream the set as CSV, JSON, or SQL inserts instead of the JSON
        envelope (F-10.5).
      parameters:
        - $ref: "#/components/parameters/ProjectID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateTestDataRequest"
      responses:
        "200":
          description: The generated set, or the export stream when a format was named.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestDataSet"
            text/csv:
              schema:
                type: string
            application/sql:
              schema:
                type: string
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/BadRequest" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/projects/{projectID}/test-data/invalid:
    post:
      operationId: generateInvalidTestData
      tags: [testdata]
      summary: Generate boundary and invalid records
      description: |
        Every record breaks exactly **one** constraint and names which (F-10.3). One at
        a time, because a payload that is simultaneously too long, the wrong type, and
        missing a required field tells you the endpoint rejected something without
        telling you what it checks.

        The values come from the schema's own constraints — a field with `maxLength: 40`
        gets a 41-character value, a field with `minimum: 1` gets 0 — so the set is
        specific to the API rather than generic noise. `injection` cases carry
        attack-shaped values; the expectation for those is not "rejected" but "stored
        and returned verbatim, never executed", because an API may legitimately accept
        an apostrophe in a surname.
      parameters:
        - $ref: "#/components/parameters/ProjectID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateInvalidTestDataRequest"
      responses:
        "200":
          description: The invalid set, each record naming the constraint it breaks.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvalidTestDataSet"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/BadRequest" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/projects/{projectID}/test-data/tables:
    get:
      operationId: listDumpTables
      tags: [testdata]
      summary: What tables an uploaded SQL dump declares
      description: |
        Reads `CREATE TABLE` statements out of a schema-only dump and reports the columns
        in the terms that shape generated data: type, length, nullability, and any value
        list a check constraint declares (F-3.7).

        The parse is deliberately narrow — triggers, functions, grants, and data are
        ignored. A dump is a program, and a parser with opinions about a client's schema
        beyond its column list would be a parser it cannot justify.
      parameters:
        - $ref: "#/components/parameters/ProjectID"
        - name: artifactID
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: The tables the dump declares.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DumpTableList"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422":
          description: The artifact is not a readable schema dump.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500": { $ref: "#/components/responses/InternalError" }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: qavia_session
      description: |
        Server-side session. Sessions live in Postgres rather than in a JWT, so
        revocation is immediate: an offboarded employee's access ends when the row
        is deleted, with no token TTL to wait out.
  parameters:
...
    ProjectID:
      name: projectID
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
    Error:
      type: object
      description: |
        The single error shape. Produced only by the mapper in
        internal/platform/httpx, so every failure looks the same.
      required: [code, message]
      properties:
        code:
          type: string
          description: |
            Stable and machine readable. Branch on this, never on message. Every
            value is catalogued in docs/error-codes.md.
          example: target_host_not_allowed
        message:
          type: string
          description: Safe to show a user, and says what to do next.
          example: Host "api.example.test" is not on this project's allowlist. Add it in Settings, Project, Targets.
        details:
          type: object
          additionalProperties: true
          description: Field-level context for inline rendering. Never carries a secret.
          example:
            host: api.example.test
        incidentId:
          type: string
          description: |
            Present only on a 500. The handle between what the user saw and the
            full wrapped cause in the log, which is never returned.
          example: 9f2c1a7e
...
    DumpColumn:
      type: object
      required: [name, type, notNull]
      properties:
        name:
          type: string
        type:
          type: string
          description: The declared SQL type without its length, lowercased.
        length:
          type: integer
          description: The character limit where the type declared one, so a generated value fits the column.
        notNull:
          type: boolean
        primaryKey:
          type: boolean
        enum:
          type: array
          description: The value list from a check constraint, which is how a dump usually expresses an enumeration.
          items:
            type: string
    DumpTable:
      type: object
      required: [name, columns]
      properties:
        name:
          type: string
        columns:
          type: array
          items:
            $ref: "#/components/schemas/DumpColumn"
    DumpTableList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/DumpTable"
        warnings:
          type: array
          description: What the parse could not read, named rather than silently skipped.
          items:
            type: string
...
    TestDataFormat:
      type: string
      description: Absent returns the JSON envelope with the seed; a value streams the export.
      enum: [json, csv, sql]
    TestDataSource:
      type: object
      description: >-
        Where the shape comes from: an endpoint of the project's specification, or a
        table of an uploaded SQL dump. Exactly one is needed.
      properties:
        dumpArtifactId:
          type: string
          format: uuid
          description: >-
            An uploaded `sql_dump` artifact. Parsed locally: no database credential and
            no MCP server are involved, which is why this path works on a fresh install.
        table:
          type: string
          maxLength: 120
          description: Which table of the dump. Optional when the dump declares exactly one.
        endpoint:
          type: string
          description: "\"METHOD /path\" from this project's specification. The request body's schema is used, or the success response's when there is no body."
          maxLength: 500
        response:
          type: string
          description: Generate from this response's schema instead of the request's, for seeding a fixture or a mock.
          maxLength: 10
    GenerateTestDataRequest:
      type: object
      required: [source]
      properties:
        source:
          $ref: "#/components/schemas/TestDataSource"
        count:
          type: integer
          minimum: 1
          maximum: 10000
          default: 10
        seed:
          type: integer
          format: int64
          description: >-
            Reproduces an earlier set exactly. Omit to be given one, which the response
            reports so the same data can be asked for again.
        locale:
          type: string
          maxLength: 20
          description: Shapes the values that have a country, such as an address or a phone number.
        fields:
          type: array
          maxItems: 20
          description: >-
            Fields to fill with an AI pass. Opt-in per field: the bulk is a seeded
            faker, and anything the model does not answer keeps the generated value.
          items:
            type: string
        format:
          $ref: "#/components/schemas/TestDataFormat"
        table:
          type: string
          maxLength: 80
          description: Table name for SQL output. Filtered to a safe identifier, not merely quoted.
    GenerateInvalidTestDataRequest:
      type: object
      required: [source]
      properties:
        source:
          $ref: "#/components/schemas/TestDataSource"
        kinds:
          type: array
          maxItems: 20
          description: Which families of violation. Absent means every kind the schema supports.
          items:
            $ref: "#/components/schemas/InvalidDataKind"
        limit:
          type: integer
          minimum: 1
          maximum: 500
          default: 50
        seed:
          type: integer
          format: int64
        locale:
          type: string
          maxLength: 20
    InvalidDataKind:
      type: string
      enum: [missing, empty, type, too_long, too_short, too_large, too_small, enum, format, pattern, injection, declined_card]
    TestDataSet:
      type: object
      required: [seed, records]
      properties:
        seed:
          type: integer
          format: int64
          description: The seed that produced this set. The same seed and count reproduce it exactly.
        records:
          type: array
          items:
            type: object
            additionalProperties: true
        warnings:
          type: array
          description: Schemas the walk could not do much with, and what the AI pass did or did not fill.
          items:
            type: string
    InvalidTestDataRecord:
      type: object
      required: [field, kind, record]
      properties:
        field:
          type: string
        kind:
          $ref: "#/components/schemas/InvalidDataKind"
        violates:
          type: string
          description: The constraint this record breaks, in words.
        expectation:
          type: string
          description: What the endpoint should do about it. This is the assertion somebody writes from the record.
        record:
          type: object
          additionalProperties: true
    InvalidTestDataSet:
      type: object
      required: [seed, records]
      properties:
        seed:
          type: integer
          format: int64
        records:
          type: array
          items:
            $ref: "#/components/schemas/InvalidTestDataRecord"
        skipped:
          type: array
          description: >-
            Fields that produced no case, so an unconstrained field is distinguishable
            from one the generator ignored.
          items:
            type: string
    TestDataShape:
      type: object
      required: [endpoint, hasRequest]
      properties:
        endpoint:
          type: string
        summary:
          type: string
        hasRequest:
          type: boolean
        responses:
          type: array
          items:
            type: string
        fields:
          type: array
          items:
            type: string
    TestDataShapeList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/TestDataShape"
...
  responses:
    BadRequest:
      description: The request is structurally valid but not permitted.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: validation_failed
            message: Email is not valid.
            details:
              field: email
    Unauthorized:
      description: No valid session.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: unauthenticated
            message: Sign in to continue.
    Forbidden:
      description: Authenticated, but not permitted.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: role_required
            message: This action requires the Admin role.
    NotFound:
      description: |
        Not found, or not visible to this caller. The two are deliberately
        indistinguishable.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: project_not_found
            message: Project not found.
...
    InternalError:
      description: |
        An unexpected failure. Carries an incident ID and nothing else; the cause
        is in the structured log against that ID.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: internal_error
            message: Something went wrong on our side.
            incidentId: 9f2c1a7e